"Add scroll event listeners to header and scrollup components for scroll effects"
This diff adds scroll event listeners to the Header and Scrollup components. In the Header component, when the user scrolls down the page, the header will have a "show-header" class added to it, and when the user scrolls back up, the class will be removed. This class changes the header's background color to make it more visible as the user scrolls.

In the Scrollup component, when the user scrolls down the page, the scrollup button will have a "show-scroll" class added to it, and when the user scrolls back up, the class will be removed. This class changes the button's position from bottom:-20% to bottom:3rem, making it visible to the user when they scroll down the page.

The diff also includes changes to the header.css and scrollup.css files to style the new classes.